Renamed ShiftProjectBelow to CollapseShifts
