+++ connector/src/main/scala/quasar/qscript/qsu/MinimizeAutoJoins.scala
@@ -51,8 +51,7 @@ final class MinimizeAutoJoins[T[_[_]]: BirecursiveT: EqualT: ShowT: RenderTreeT]
 
   private val Minimizers = List(
     minimizers.MergeReductions[T],
-    // minimizers.CollapseShifts[T],
-    minimizers.ShiftProjectBelow[T])
+    minimizers.CollapseShifts[T])
 
   private val func = construction.Func[T]
   private val srcHole: Hole = SrcHole   // wtb smart constructor
